requirements.txt
absl-py==0.7.0
astor==0.7.1
backcall==0.1.0
colorama==0.4.1
cycler==0.10.0
decorator==4.4.0
Django==2.1.7
django-crispy-forms==1.7.2
django-debug-toolbar==1.11
gast==0.2.2
grpcio==1.18.0
h5py==2.9.0
ipdb==0.12
ipykernel==5.1.0
ipython==7.4.0
ipython-genutils==0.2.0
jedi==0.13.3
jupyter-client==5.2.4
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
Markdown==3.0.1
matplotlib==3.0.2
mysql-connector-python==8.0.15
mysqlclient==1.4.2.post1
numpy==1.16.1
opencv-contrib-python==4.0.0.21
pandas==0.24.1
parso==0.3.4
pickleshare==0.7.5
Pillow==5.4.1
prompt-toolkit==2.0.9
protobuf==3.6.1
Pygments==2.3.1
pyparsing==2.3.1
python-dateutil==2.8.0
python-utils==2.3.0
pytz==2018.9
PyYAML==3.13
pyzmq==18.0.1
scikit-learn==0.20.3
scipy==1.2.1
six==1.12.0
sqlparse==0.3.0
tensorboard==1.12.2
tensorflow==1.12.0
termcolor==1.1.0
tornado==6.0.1
traitlets==4.3.2
utils==0.9.0
wcwidth==0.1.7
Werkzeug==0.14.1
xlwt==1.3.0
